+++ framework/src/main/java/com/github/dbmdz/flusswerk/framework/rabbitmq/ChannelCommand.java
@@ -1,0 +1,7 @@
+package com.github.dbmdz.flusswerk.framework.rabbitmq;
+
+import java.io.IOException;
+
+public interface ChannelCommand<T> {
+  T execute() throws IOException;
+}