+++ framework/src/main/java/com/github/dbmdz/flusswerk/framework/rabbitmq/ChannelListener.java
@@ -1,0 +1,6 @@
+package com.github.dbmdz.flusswerk.framework.rabbitmq;
+
+/** A ChannelListener receives notifications about channel recovery. */
+public interface ChannelListener {
+  void handleReset();
+}