+++ framework/src/main/java/com/github/dbmdz/flusswerk/framework/rabbitmq/Queue.java
@@ -4,8 +4,6 @@ import static java.util.Objects.requireNonNull;
 
 import com.github.dbmdz.flusswerk.framework.exceptions.InvalidMessageException;
 import com.github.dbmdz.flusswerk.framework.model.Message;
-import com.rabbitmq.client.Channel;
-import java.io.IOException;
 import java.util.Objects;
 import java.util.Optional;
 import org.slf4j.Logger;
@@ -17,12 +15,10 @@ public class Queue {
   private static final Logger LOGGER = LoggerFactory.getLogger(Queue.class);
 
   private final String name;
-  private final Channel channel;
   private final RabbitClient rabbitClient;
 
   Queue(String name, RabbitClient rabbitClient) {
     this.name = requireNonNull(name);
-    this.channel = rabbitClient.getChannel();
     this.rabbitClient = rabbitClient;
   }
 
@@ -30,10 +26,9 @@ public class Queue {
    * Removes all messages from a queue. These messages cannot be restored.
    *
    * @return the number of deleted messages
-   * @throws IOException if an error occurs while purging
    */
-  public int purge() throws IOException {
-    var purgeOk = channel.queuePurge(this.name);
+  public int purge() {
+    var purgeOk = rabbitClient.queuePurge(this.name);
     var deletedMessages = purgeOk.getMessageCount();
     LOGGER.warn("Purged queue {} ({} messages deleted)", this.name, deletedMessages);
     return deletedMessages;
@@ -41,10 +36,9 @@ public class Queue {
 
   /**
    * @return the number of messages in this queue.
-   * @throws IOException if communication with RabbitMQ fails.
    */
-  public long messageCount() throws IOException {
-    return channel.messageCount(this.name);
+  public long messageCount() {
+    return rabbitClient.getMessageCount(this.name);
   }
 
   /**