+++ integration-tests/src/test/java/com/github/dbmdz/flusswerk/integration/processing/SkipProcessingTest.java
@@ -54,7 +54,7 @@ import org.testcontainers.junit.jupiter.Testcontainers;
 public class SkipProcessingTest {
 
   @Container
-  static RabbitMQContainer rabbitMQContainer =
+  static final RabbitMQContainer rabbitMQContainer =
       new RabbitMQContainer("rabbitmq:3-management-alpine");
 
   private final Engine engine;
@@ -118,7 +118,7 @@ public class SkipProcessingTest {
   }
 
   @AfterEach
-  void stopEngine() throws IOException {
+  void stopEngine() {
     engine.stop();
     rabbitUtil.purgeQueues();
   }